version ticker fix
can't commit and sign off on changes to multiple files when using the GitHub editor. cannot use vs code because I am on my phone. therefore, when I needed to make an edit to a bugged icon in the personnel page, I was unable to simultaneously increment the version ticker on the login page. resolved

Signed-off-by: pumpjack <aetsofficialact@gmail.com>